feat (jwt): implement all endpoint to check bearer token

+++ app/controllers/bookmark.controller.js
@@ -1,10 +1,11 @@
-import { prisma } from "../lib/dbConnect.js";
-export * as bookmarkController from "./bookmark.controller.js";
+import { prisma } from "../lib/dbConnect.js"
+export * as bookmarkController from "./bookmark.controller.js"
+import { verifyToken } from "../lib/tokenHandler.js"
 
 // Create bookmark record
 export const createBookmark = async (req, res, next) => {
   try {
-    const { video_id, user_id, is_bookmark } = req.body;
+    const { video_id, user_id, is_bookmark } = req.body
 
     const newBookmark = await prisma.bookmarks.create({
       data: {
@@ -12,52 +13,61 @@ export const createBookmark = async (req, res, next) => {
         user_id,
         is_bookmark,
       },
-    });
+    })
 
-    res.status(201).json({ success: true, data: newBookmark });
+    res.status(201).json({ success: true, data: newBookmark })
   } catch (error) {
-    next(error);
+    next(error)
   }
-};
+}
 
 // Get all bookmarks
 export const getAllBookmarks = async (req, res, next) => {
   try {
-    const bookmarks = await prisma.bookmarks.findMany();
+    const data = verifyToken(req.headers.access_token)
+    if (data?.status) return res.status(data.status).json(data)
 
-    res.json({ success: true, data: bookmarks });
+    const bookmarks = await prisma.bookmarks.findMany()
+
+    res.json({ success: true, data: bookmarks })
   } catch (error) {
-    next(error);
+    next(error)
   }
-};
+}
 
 // Get bookmark by ID
 export const getBookmarkById = async (req, res, next) => {
   try {
-    const bookmarkId = parseInt(req.params.id);
+    const data = verifyToken(req.headers.access_token)
+    if (data?.status) return res.status(data.status).json(data)
+
+    const bookmarkId = parseInt(req.params.id)
 
     const bookmark = await prisma.bookmarks.findUnique({
       where: { id: bookmarkId },
-    });
+    })
 
     if (!bookmark) {
       return res.status(404).json({
         status: 404,
         message: "Bookmark not found.",
-      });
+      })
     }
 
-    res.json({ success: true, data: bookmark });
+    res.json({ success: true, data: bookmark })
   } catch (error) {
-    next(error);
+    next(error)
   }
-};
+}
 
 // Update bookmark by ID
 export const updateBookmark = async (req, res, next) => {
   try {
-    const bookmarkId = parseInt(req.params.id);
-    const { video_id, user_id, is_bookmark } = req.body;
+    const data = verifyToken(req.headers.access_token)
+    if (data?.status) return res.status(data.status).json(data)
+
+    const bookmarkId = parseInt(req.params.id)
+    const { video_id, user_id, is_bookmark } = req.body
 
     const updatedBookmark = await prisma.bookmarks.update({
       where: { id: bookmarkId },
@@ -66,25 +76,28 @@ export const updateBookmark = async (req, res, next) => {
         user_id,
         is_bookmark,
       },
-    });
+    })
 
-    res.json({ success: true, data: updatedBookmark });
+    res.json({ success: true, data: updatedBookmark })
   } catch (error) {
-    next(error);
+    next(error)
   }
-};
+}
 
 // Delete bookmark by ID
 export const deleteBookmark = async (req, res, next) => {
   try {
-    const bookmarkId = parseInt(req.params.id);
+    const data = verifyToken(req.headers.access_token)
+    if (data?.status) return res.status(data.status).json(data)
+
+    const bookmarkId = parseInt(req.params.id)
 
     const deletedBookmark = await prisma.bookmarks.delete({
       where: { id: bookmarkId },
-    });
+    })
 
-    res.json({ success: true, data: deletedBookmark });
+    res.json({ success: true, data: deletedBookmark })
   } catch (error) {
-    next(error);
+    next(error)
   }
-};
+}
